requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
beautifulsoup4
lxml
python-dotenv